add foler to contain all php files
