+++ implementation/reset_process.php
